pyproject.toml
[project]
name = "zadanie1"
version = "0.1.0"
description = "Zadanie 1"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
